refactor(utils): share node sourcemap extraction
Export extractSourcemapFromFile from @vitest/utils/source-map-node and reuse it in browser and vitest node paths to remove duplicated sourcemap parsing logic.
