+++ packages/utils/package.json
@@ -56,6 +56,10 @@
       "types": "./dist/source-map.d.ts",
       "default": "./dist/source-map.js"
     },
+    "./source-map-node": {
+      "types": "./dist/source-map-node.d.ts",
+      "default": "./dist/source-map-node.js"
+    },
     "./serialize": {
       "types": "./dist/serialize.d.ts",
       "default": "./dist/serialize.js"
@@ -82,10 +86,12 @@
   },
   "dependencies": {
     "@vitest/pretty-format": "workspace:*",
+    "convert-source-map": "^2.0.0",
     "tinyrainbow": "catalog:"
   },
   "devDependencies": {
     "@jridgewell/trace-mapping": "catalog:",
+    "@types/convert-source-map": "^2.0.3",
     "@types/estree": "catalog:",
     "diff-sequences": "^29.6.3",
     "loupe": "^3.2.1"
