+++ packages/vitest/src/node/test-run.ts
@@ -7,7 +7,6 @@ import type {
 } from '@vitest/runner'
 import type { TaskEventData, TestArtifact } from '@vitest/runner/types/tasks'
 import type { SerializedError } from '@vitest/utils'
-import type { SourceMap } from 'rollup'
 import type { UserConsoleLog } from '../types/general'
 import type { Vitest } from './core'
 import type { TestProject } from './project'
@@ -18,11 +17,10 @@ import assert from 'node:assert'
 import { createHash } from 'node:crypto'
 import { existsSync, readFileSync } from 'node:fs'
 import { copyFile, mkdir, writeFile } from 'node:fs/promises'
-import path from 'node:path'
 import { isPrimitive } from '@vitest/utils/helpers'
 import { serializeValue } from '@vitest/utils/serialize'
 import { parseErrorStacktrace } from '@vitest/utils/source-map'
-import convertSourceMap from 'convert-source-map'
+import { extractSourcemapFromFile } from '@vitest/utils/source-map-node'
 import mime from 'mime/lite'
 import { basename, extname, resolve } from 'pathe'
 
@@ -182,7 +180,7 @@ export class TestRun {
               if (!mod?.transformResult && existsSync(file)) {
                 const code = readFileSync(file, 'utf-8')
                 const result = extractSourcemapFromFile(code, file)
-                return result
+                return result?.map
               }
             },
           })
@@ -312,32 +310,4 @@ export class TestRun {
 function sanitizeFilePath(s: string): string {
   // eslint-disable-next-line no-control-regex
   return s.replace(/[\x00-\x2C\x2E\x2F\x3A-\x40\x5B-\x60\x7B-\x7F]+/g, '-')
-}
-
-// based on vite
-// https://github.com/vitejs/vite/blob/84079a84ad94de4c1ef4f1bdb2ab448ff2c01196/packages/vite/src/node/server/sourcemap.ts#L149
-function extractSourcemapFromFile(
-  code: string,
-  filePath: string,
-): SourceMap | undefined {
-  const map = (
-    convertSourceMap.fromSource(code)
-    || (convertSourceMap.fromMapFileSource(
-      code,
-      createConvertSourceMapReadMap(filePath),
-    ))
-  )?.toObject()
-  return map
-}
-
-function createConvertSourceMapReadMap(originalFileName: string) {
-  return (filename: string) => {
-    // convertSourceMap can detect invalid filename from comments.
-    // fallback to empty source map to avoid errors.
-    const targetPath = path.resolve(path.dirname(originalFileName), filename)
-    if (existsSync(targetPath)) {
-      return readFileSync(targetPath, 'utf-8')
-    }
-    return '{}'
-  }
 }
